Search objects page with initial page without results
